requirements.txt
certifi==2019.9.11
JayDeBeApi==1.1.1
JPype1==0.6.3
numpy==1.17.3
PyYAML==5.1.2
wincertstore==0.2
